fhdl.verilog: declare outputs from specials, but assigned in always(*) blocks, as reg instead of wire
Fixes #119

Currently, the migen code assumes that all outputs from specials must be
wire, which is true logically. But wires can also be assigned inside
always(*) blocks, which then should be declared as regs.

Solution:
We find out all the output signals from specials (set special_outs_only).
Then we make a list which has comb statements grouped-by-target signals.
We look whether signals in this list are present in `special_outs_only`
set or not. If they are , we remove those signals from the wires set.

Signed-off-by: Rohit Singh <rohit91.2008@gmail.com>